% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{checkNumeric}
\alias{checkNumeric}
\title{Wrapper for \code{is.numeric}}
\usage{
checkNumeric(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Breaks function and returns error message if x
is non-numeric
}
\description{
Wrapper - check if vector is numeric
}

