% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CormorantData}
\alias{CormorantData}
\title{O2 consumption data -- flow-through respirometry for cormorants}
\format{A data frame with 447 rows and 2 variables:
\itemize{
		\item Time: time in hours
		\item VO2.ml.min: Rate of oxygen consumption (mL O2/min)
}}
\description{
A dataset containing O2 consumption data for a single Great cormorant. This
dataset represents a small example of data used for analyses described in White
et al. 2011. Metabolic rate throughout the annual cycle reveals the demands
of an Arctic existence in Great Cormorants. Ecology 92: 475--486.
}

