% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{locReg}
\alias{locReg}
\title{Perform local regression for specified window}
\usage{
locReg(wins, xall, yall, resids = FALSE)
}
\arguments{
\item{wins}{A matrix of vector positions, where wins[,1] contains left boundary positions,
and wins[,2] contains right boundary positions for the desired local regressions. Default
behaviour is optimized for use with \code{\link{getWindows}} in \code{\link{rankLocReg}},
but can accept user-defined windows.}

\item{xall}{A numeric vector.}

\item{yall}{A numeric vector.}

\item{resids}{Logical. Are residuals to be returned?}
}
\value{
A data frame if \code{resids} is \code{FALSE}. A list if \code{resids} is \code{TRUE}.
}
\description{
Local Regression Function.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
# rank L metric by method 'eq'
wins  <-  getWindows(x=TestO2data$D, alpha=0.3)
reg   <-  locReg(wins[1, ], xall=TestO2data$time, yall=TestO2data$D, resids=TRUE)
}
\author{
Colin Olito and Diego Barneche.
}
\seealso{
\code{\link{getWindows}}, \code{\link{rankLocReg}}.
}

