% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{outputRankLocRegPlot}
\alias{outputRankLocRegPlot}
\title{Plotting 25 best local linear regressions}
\usage{
outputRankLocRegPlot(allRegs)
}
\arguments{
\item{allRegs}{An object of class \code{rankLocReg}}
}
\value{
A 25-plot panel
}
\description{
Plotting 25 best local linear regressions
}
\details{
Generates scatterplots for 25 best local regressions
}
\seealso{
\code{rankLocReg.default}
}

