% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pcRank}
\alias{pcRank}
\title{Calculate the percentile values of a vector x}
\usage{
pcRank(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of percentiles truncated between 0 and 1
}
\description{
Calculate the percentile values of a vector x
}

