% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PelagicData}
\alias{PelagicData}
\title{O2 production data for pelagic phytoplankton in ponds}
\format{A data frame with 242 rows and 5 variables:
\itemize{
		\item pond: pond ID
		\item month.num: Sampling month ID
		\item flux: light conditions
		\item Time.m: Time in minutes
		\item O2: O2 (mg/L) data for phytoplankton
}}
\description{
A dataset containing O2 production time series for
pelagic phytoplankton in several ponds and light conditions.
These data represent a small example of the data used for analyses
described in Yvon-Durocher et al. (2015). Five years of experimental
warming increases the biodiversity and productivity of phytoplankton.
PLoS Biology 13(12): e1002324. doi:10.1371/journal. pbio.1002324.
}

