% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.rankLocReg}
\alias{plot.rankLocReg}
\title{Plotting chosen local linear regression}
\usage{
\method{plot}{rankLocReg}(allRegs, rank = 1)
}
\arguments{
\item{allRegs}{An object of class \code{rankLocReg}}

\item{rank}{Position, as in row number from input \code{allRegs}, of local regression to be plotted.}
}
\value{
A 5-plot panel
}
\description{
Plotting chosen local linear regression
}
\details{
Generates a scatterplot + residual-plot diagnostics for chosen local regression
}
\seealso{
\code{rankLocReg.default}
}

