% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.rankLocReg}
\alias{print.summary.rankLocReg}
\title{Wrapper summary for object of class \code{rankLocReg}}
\usage{
\method{print}{summary.rankLocReg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rankLocReg}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A summary list with main features calculated by function \code{\link{rankLocReg}}.
}
\description{
Wrapper summary for object of class \code{rankLocReg}.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
# rank L metric by method 'eq'
allRegs  <-  rankLocReg(xall=TestO2data$time, yall=TestO2data$D, alpha=0.3, method="eq", verbose=TRUE)
summary(allRegs)
}
\author{
Diego Barneche.
}
\seealso{
\code{\link{rankLocReg}}, \code{\link{summary.rankLocReg}}
}

