% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rankLocReg.default}
\alias{rankLocReg.default}
\title{Ranking local linear regressions}
\usage{
\method{rankLocReg}{default}(xall, yall, alpha, method = c("ns", "eq", "pc"),
  verbose = TRUE)
}
\arguments{
\item{xall}{A numeric vector}

\item{yall}{A numeric vector}

\item{alpha}{Window size. Needs to be higher than 0 and lower or equal to 1.}

\item{method}{Ranking method. See details.}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
A data frame with local regressions ranked by metric L following raking method chosen by argument \code{method}.
}
\description{
Ranking local linear regressions
}
\details{
To be completed.
}
\seealso{
\code{locReg}
}

