% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rankLocReg}
\alias{rankLocReg}
\title{rankLocReg}
\usage{
rankLocReg(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments to \code{rankLocReg.default}}
}
\value{
Default function \code{rankLocReg.default}
}
\description{
Wrapper - calls function that creates class rankLocReg
}
\seealso{
\code{rankLocReg.default}
}

