% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{reRank}
\alias{reRank}
\title{Re-rank \code{rankLocReg}}
\usage{
reRank(x, newMethod)
}
\arguments{
\item{x}{An object of class \code{rankLocReg}.}

\item{newMethod}{Method with which object of class \code{rankLocReg} should be re-ranked.}
}
\value{
An object of class \code{rankLocReg}.
}
\description{
Re-rank an object of class \code{rankLocReg}.
}
\details{
This function also updates \code{x$call} and \code{x$method}.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
x  <-  rankLocReg(xall=TestO2data$time, yall=TestO2data$D, alpha=0.8, method="eq", verbose=TRUE)
x2  <-  reRank(x, newMethod="pc")
# check outputs
x$call; x2$call; x$method; x2$method
}
\author{
Diego Barneche.
}

