% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{skew}
\alias{skew}
\title{Sample skewness (Fisher-Pearson Standardized Third Moment Coefficient)}
\usage{
skew(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of length 1
}
\description{
Sample skewness
}
\details{
This function is a dependency for \code{findLocLin}
where it is used to calculate the (sample) skewness of standardized residuals.
}

