% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{stripNAs}
\alias{stripNAs}
\title{Remove NA}
\usage{
stripNAs(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A data frame with complete cases.
}
\description{
Wrapper - strips NA from input x and y.
}

