% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.rankLocReg}
\alias{summary.rankLocReg}
\title{Summary for object of class \code{rankLocReg}}
\usage{
\method{summary}{rankLocReg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rankLocReg}}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A summary list with main features calculated by function \code{rankLocReg}
}
\description{
Summary for object of class \code{rankLocReg}
}
\seealso{
\code{rankLocReg}
}

