% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestO2data}
\alias{TestO2data}
\title{O2 consumption data for sea urchins}
\format{A data frame with 166 rows and 5 variables:
\itemize{
		\item time: time in minutes
		\item A: Volume O2 (mL) data for urchin A
		\item B: Volume O2 (mL) data for urchin B
		\item C: Volume O2 (mL) data for urchin C
		\item D: Volume O2 (mL) data for urchin D
}}
\description{
A dataset containing VO2 time series for 4 individual sea urchins.
Data are provided by Colin Olito (unpublished data).
}

