% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{thinData}
\alias{thinData}
\title{Thin a large data set}
\usage{
thinData(xall, yall, by = 2)
}
\arguments{
\item{xall}{A numeric vector.}

\item{yall}{A numeric vector.}

\item{by}{An integer.}
}
\value{
A list of data frames with thinned observations.
}
\description{
Thin large data sets to contain fewer observations
}
\author{
Colin Olito.
}

