% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{transparentColor}
\alias{transparentColor}
\title{Creates transparent colours}
\usage{
transparentColor(col, opacity = 0.5)
}
\arguments{
\item{col}{Colour}

\item{opacity}{Relative y-axis position (in proportion) where character is to be plotted}
}
\description{
Creates transparent colours
}
\author{
Richard G. FitzJohn.
}

