% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{checkEqualLength}
\alias{checkEqualLength}
\title{Check for equal lengths}
\usage{
checkEqualLength(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
Logical.
}
\description{
Wrapper - Checks that lengths of inputs x and y are equal.
}
\author{
Colin Olito and Diego Barneche.
}

