% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{locReg}
\alias{locReg}
\title{To be written}
\usage{
locReg(wins, xall, yall, resids = FALSE)
}
\arguments{
\item{wins}{A matrix of vector positions obtained from \code{\link{getWindows}}.}

\item{xall}{A numeric vector.}

\item{yall}{A numeric vector.}

\item{resids}{Logical. Are residuals to be returned?}
}
\value{
A data frame if \code{resids} is \code{FALSE}. A list if \code{resids} is \code{TRUE}.
}
\description{
To be written.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
# rank L metric by method 'eq'
wins  <-  getWindows(x=TestO2data$D, alpha=0.3)
reg   <-  locReg(wins[1, ], xall=TestO2data$time, yall=TestO2data$D, resids=TRUE)
}
\seealso{
\code{\link{getWindows}}.
}

