% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{outputRankLocRegPlot}
\alias{outputRankLocRegPlot}
\title{Plotting 25 best local linear regressions}
\usage{
outputRankLocRegPlot(allRegs)
}
\arguments{
\item{allRegs}{An object of class \code{rankLocReg}.}
}
\value{
Generates scatterplots for 25 best local regressions.
}
\description{
Plotting 25 best local linear regressions.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
# rank L metric by method 'eq'
allRegs  <-  rankLocReg(xall=TestO2data$time, yall=TestO2data$D, alpha=0.3, method="eq", verbose=TRUE)
outputRankLocRegPlot(allRegs)
}
\author{
Colin Olito and Diego Barneche.
}
\seealso{
\code{rankLocReg.default}
}

