% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.rankLocReg}
\alias{plot.rankLocReg}
\title{Plotting chosen local linear regression}
\usage{
\method{plot}{rankLocReg}(x, ..., rank = 1)
}
\arguments{
\item{x}{An object of class \code{rankLocReg}.}

\item{...}{Other parameters to be passed through to plotting functions.}

\item{rank}{Position, as in row number from input \code{x}, of local regression to be plotted.}
}
\value{
Generates a distribution of all local regression slopes + residual-plot diagnostics for chosen local regression + scatterplot.
}
\description{
Plotting chosen local linear regression.
}
\examples{
# load sea urchin respirometry data
data(UrchinData)
# rank L metric by method 'eq'
allRegs  <-  rankLocReg(xall=UrchinData$time, yall=UrchinData$D, alpha=0.3, method="eq", verbose=TRUE)
# plot best local regression
plot(allRegs, rank=1)
}
\author{
Colin Olito and Diego Barneche.
}
\seealso{
\code{\link{rankLocReg.default}}.
}

