% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotBeta1}
\alias{plotBeta1}
\title{Distribution of all local slopes}
\usage{
plotBeta1(allRegs)
}
\arguments{
\item{allRegs}{An object of class \code{rankLocReg}}
}
\description{
Distribution of all local slopes
}
\details{
Generates a distribution of all local regression slopes
}
\seealso{
\code{rankLocReg.default}
}

