% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.rankLocReg}
\alias{print.summary.rankLocReg}
\title{Wrapper summary for object of class \code{rankLocReg}}
\usage{
\method{print}{summary.rankLocReg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rankLocReg}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A summary list with main features calculated by function \code{rankLocReg}
}
\description{
Wrapper summary for object of class \code{rankLocReg}
}
\seealso{
\code{rankLocReg}, \code{summary.rankLocReg}
}

