% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rankLocReg}
\alias{rankLocReg}
\title{Ranking local linear regressions}
\usage{
rankLocReg(xall, yall, alpha, method = c("ns", "eq", "pc"), verbose = TRUE)
}
\arguments{
\item{xall}{A numeric vector.}

\item{yall}{A numeric vector.}

\item{alpha}{Window size. Needs to be higher than 0 and lower or equal to 1.}

\item{method}{Ranking method. See details.}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
Default function \code{\link{rankLocReg.default}}.
}
\description{
Wrapper - calls function that creates class \code{rankLocReg}.
}
\details{
If method is unspecified, default to \code{ns}.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
# rank L metric by method 'eq'
allRegs  <-  rankLocReg(xall=TestO2data$time, yall=TestO2data$D, alpha=0.3, method="eq", verbose=TRUE)
}
\seealso{
\code{\link{rankLocReg.default}}.
}

