% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rankLocReg}
\alias{rankLocReg}
\title{Ranking local linear regressions according to linearity}
\usage{
rankLocReg(xall, yall, alpha, method = c("ns", "eq", "pc"), verbose = TRUE)
}
\arguments{
\item{xall}{A numeric vector.}

\item{yall}{A numeric vector.}

\item{alpha}{Minimum window size, expressed as a proportion of total data. Must be higher than 0 and less-than or equal to 1.}

\item{method}{Ranking method. See details.}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
A data frame with important output from all local regressions, ranked by metric \code{L} following raking method chosen by argument \code{method}.
}
\description{
Wrapper - calls function that creates class \code{rankLocReg}.
}
\details{
\code{rankLocReg} is the main function around which \pkg{LoLinR} is built. Given independent and dependent variables
for a time series or trace data set, \code{rankLocReg} tries to find the 'most linear' ordered subset of the full data set.
This is accomplished by fitting all possible local linear regressions with minimum window size \code{alpha}, and ranking them
according to the combined linearity metric $L$. $L$ quantifies linearity from 1) the skewness of the standardized residuals,
2) the range of the 95% confidence interval around the regression slope $\beta_1$, and 3) auto-correlation among the standardized
residuals (a modified Breusch-Godfrey $R^2$). These three components of $L$ can be weighted in 3 different ways: unweighted (\code{method="ns"}),
equal weights (\code{method="eq"}), and percentile ranks (\code{method="pc"}). If method is unspecified, default to \code{ns}.
For highly skewed, or otherwise ill-behaved data we strongly advise examining the relative behaviour of the different weighting
methods using \code{\link{plotBeta1}}.

For data sets with greater tha ~500 observations, we suggest thinning the number of observations using \code{\link{thinData}},
given that this does not compromise the resolution of the data for the question of interest.
}
\examples{
# load sea urchin respirometry data
data(UrchinData)
# rank L metric by method 'eq'
allRegs  <-  rankLocReg(xall=UrchinData$time, yall=UrchinData$D, alpha=0.3, method="eq", verbose=TRUE)
}
\author{
Colin Olito and Diego Barneche.
}
\seealso{
\code{\link{locReg}}, \code{\link{thinData}}, \code{\link{plotBeta1}}.
}

