% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rounded}
\alias{rounded}
\title{Rounded numbers for plotting}
\usage{
rounded(value, precision = 1)
}
\arguments{
\item{value}{A numeric vector.}

\item{precision}{Number of rounding digits.}
}
\value{
A character vector.
}
\description{
Internal. Create nice rounded numbers for plotting.
}

