% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.rankLocReg}
\alias{summary.rankLocReg}
\title{Summary for object of class \code{rankLocReg}}
\usage{
\method{summary}{rankLocReg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rankLocReg}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
A summary list with main features calculated by function \code{\link{rankLocReg}}.
}
\description{
Summary for object of class \code{rankLocReg}.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
# rank L metric by method 'eq'
allRegs  <-  rankLocReg(xall=TestO2data$time, yall=TestO2data$D, alpha=0.3, method="eq", verbose=TRUE)
summary(allRegs)
}
\seealso{
\code{\link{rankLocReg}}.
}

