% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{thinData}
\alias{thinData}
\title{Thin a large data set}
\usage{
thinData(data, xy = c(1, 2), by = 2)
}
\arguments{
\item{data}{A data.frame.}

\item{xy}{A numeric vector designating the columns where x and y are located in data.}

\item{by}{An integer specifying the thinning rate. Analogous to the \code{by} argument
from the \code{seq} command from the \pkg{base} package.}
}
\value{
A list of data frames with thinned observations.
}
\description{
Thin large data sets to contain fewer observations
}
\details{
This function is used do reduce the size of large data sets to improve
overall computation time for \code{\link{rankLocReg}}. \code{\link{locReg}} takes
approximately 5 minutes to complete for data sets of ~500 observations, and
computation time increases exponentially with larger data sets. Provided that the
data is of suitably high resolution, thinning larger data sets is a reasonable
solution for reducing this computation time.
}
\examples{
# load sea urchin respirometry data
data(TestO2data)
head(TestO2data)
# thin data for individual "D" by 1/2
newData  <-  thinData(TestO2data, xy=c(1,5), by=2)
head(newData[[1]])
}
\author{
Colin Olito.
#'
}

