% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{thinData}
\alias{thinData}
\title{Thin a large data set}
\usage{
thinData(data, xy = c(1, 2), by = 2)
}
\arguments{
\item{data}{A data.frame.}

\item{xy}{A numeric vector designating the columns where x and y are located in data.}

\item{by}{An integer specifying the thinning rate.}
}
\value{
A list of data frames with thinned observations.
}
\description{
Thin large data sets to contain fewer observations
}
\author{
Colin Olito.
}

