% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{whiteGrid}
\alias{whiteGrid}
\title{Equally-spaced white lines on plot window}
\usage{
whiteGrid(...)
}
\arguments{
\item{...}{Additional arguments to internal function \code{proportionalLabel}}
}
\description{
Draw equally-spaced white lines on plot window
}

